requirements.txt
fastapi
uvicorn
requests
beautifulsoup4
aiohttp
pycryptodome